https://jira.catena-x.net/browse/A1ODT-127 move routing to app component
