+++ devops-tooling/dashboard/src/App.test.tsx
@@ -1,9 +1,16 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import App from './App';
+import { BrowserRouter } from 'react-router-dom';
 
-test('renders Devops dashboard', () => {
+const renderWithRouter = (ui, {route = '/'} = {}) => {
+  window.history.pushState({}, 'Test page', route)
+
+  return render(ui, {wrapper: BrowserRouter})
+}
+
+test('renders Login by default', () => {
   render(<App />);
-  const linkElement = screen.getByText(/DevOps Tooling/i);
+  const linkElement = screen.getByTestId("login");
   expect(linkElement).toBeInTheDocument();
 });
