+++ devops-tooling/dashboard/src/App.tsx
@@ -1,28 +1,27 @@
 import React from 'react';
-import Dashboard from './components/Dashboard/Dashboard';
 import ThemeProvider from '@mui/system/ThemeProvider'
-import Grid from '@mui/material/Grid'
-import Button from '@mui/material/Button'
 import theme from './Theme';
-
-import useAuth from './Auth/useAuth'
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import Login from './components/pages/Login/Login';
+import RequireAuth from './Auth/RequireAuth';
+import AuthProvider from './Auth/AuthProvider';
+import Dashboard from './components/pages/Dashboard/Dashboard';
 
 function App() {
-  const auth = useAuth();
-
-  function handleClick () {
-    auth.signOut(()=>{});
-  }
-
   return (
     <ThemeProvider theme={theme}>
-      <Grid container direction="column">
-        <Grid container justifyContent="center" sx={{p: 2}}>
-          <span>DevOps Tooling</span>
-          <Button variant="contained" color="primary" onClick={handleClick}>Logout</Button>
-        </Grid>
-        <Dashboard />
-      </Grid>
+      <BrowserRouter>
+        <AuthProvider>
+          <Routes>
+            <Route path="/" element={
+              <Login />
+            } />
+            <Route path="dashboard" element={
+              <RequireAuth><Dashboard /></RequireAuth>
+            }   />
+          </Routes>
+        </AuthProvider>
+      </BrowserRouter>
     </ThemeProvider >
   );
 }
